python/generate-pdf/requirements.txt
fpdf2==2.7.6
faker==36.1.0